feature:player's color
