+++ Escape From Note/Assets/Materials/M_Yellow.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61a8873785ec6ff4c81f8e885a06b2aa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
