+++ Escape From Note/Assets/Prefabs/Chalk_Yellow.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7579f4824ff6ea945867979bdf980cbc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
