+++ Escape From Note/Assets/Scripts/GameScene/Chalk.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fc4c30440ea96b438529d8d7477da40
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
