+++ Escape From Note/Assets/Scripts/GameScene/Chalk/Chalk_Yellow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d05b313d0be752447b12f8dcad2f7ad8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
